requirements.txt
Flask
flask-cors
requests
google-api-python-client
google-auth
SQLAlchemy
python-dotenv
